refactor: extract browser metrics helpers
